Moved input logic to dedicated file
