+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "stdio.h": "c"
+        "stdio.h": "c",
+        "commands.h": "c"
     }
 }
